+++ injected/src/features/web-compat.js
@@ -126,10 +126,6 @@ export class WebCompat extends ContentFeature {
         if (this.getFeatureSettingEnabled('modifyCookies')) {
             this.modifyCookies();
         }
-
-        if (this.getFeatureSettingEnabled('videoPlayback')) {
-            this.videoPlaybackObserve();
-        }
     }
 
     /** Shim Web Share API in Android WebView */
@@ -756,57 +752,6 @@ export class WebCompat extends ContentFeature {
             });
             this.forceViewportTag(viewportTag, newContent.join(', '));
         }
-    }
-
-    videoPlaybackObserve() {
-        if (document.readyState === 'loading') {
-            // if the document is not ready, we may miss the original viewport tag
-            document.addEventListener('DOMContentLoaded', () => this.videoPlaybackObserveInner());
-        } else {
-            this.videoPlaybackObserveInner();
-        }
-    }
-
-    videoPlaybackObserveInner() {
-        const seenVideoElements = new WeakSet();
-        function addPlayObserver(video) {
-            if (seenVideoElements.has(video)) {
-                return; // already observed
-            }
-            seenVideoElements.add(video);
-            video.addEventListener('play', () => {
-                if (navigator.userActivation.isActive) {
-                    console.log('user interaction');
-                }
-            });
-        }
-
-        function addListenersToAllVideos(node) {
-            const videos = node.querySelectorAll('video');
-            videos.forEach((video) => {
-                addPlayObserver(video);
-            });
-        }
-        addListenersToAllVideos(document.body);
-        const observerCallback = (mutationsList) => {
-            for (const mutation of mutationsList) {
-                if (mutation.type === 'childList') {
-                    mutation.addedNodes.forEach((node) => {
-                        if (node.nodeType === Node.ELEMENT_NODE) {
-                            if (node.tagName === 'VIDEO') {
-                                addPlayObserver(node);
-                            }
-                            addListenersToAllVideos(node);
-                        }
-                    });
-                }
-            }
-        };
-        const observer = new MutationObserver(observerCallback);
-        observer.observe(document.body, {
-            childList: true,
-            subtree: true,
-        });
     }
 }
